chore: documentation and clean on player stuff, trying to set things up for a clean codebase
