--- Assets/Scripts/Player/Player.cs
@@ -1,26 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Player : MonoBehaviour
-{
-    public float speed;
-    public Rigidbody rb;
-
-    private void Start()
-    {
-        rb = GetComponent<Rigidbody>();
-    }
-
-    void FixedUpdate ()
-    {
-        movement();
-	}
-    void movement()
-    {
-        Vector3 input = new Vector3(Input.GetAxisRaw("Horizontal"), 0, Input.GetAxisRaw("Vertical"));
-        Vector3 direction = input.normalized;
-        Vector3 velocity = direction * speed;
-        rb.velocity = velocity;
-    }
-}
